+++ classes/MyCookie.py
@@ -1,0 +1,12 @@
+from http.cookiejar import MozillaCookieJar
+import threading
+
+class MyCookieJar(MozillaCookieJar):
+    def __getstate__(self):
+        state = self.__dict__.copy()
+        del state['_cookies_lock']
+        return state
+
+    def __setstate__(self, state):
+        self.__dict__ = state
+        self._cookies_lock = threading.RLock()
