+++ main.py
@@ -1,0 +1,24 @@
+import sys
+
+from helpers import get_listing, download, download_all
+from classes import *
+
+def main():
+    options = Options(sys.argv)
+    session = Session(options)
+    if options.clear_cookies:
+        session.logout()
+    if options.listing:
+        get_listing(session)
+    if options.download:
+        download(session)
+    if options.download_all:
+        download_all(session)
+
+
+if __name__ == '__main__':
+    try:
+        main()
+    except KeyboardInterrupt:
+        print()
+        exit(1)
