+++ variables.py
@@ -1,0 +1,1 @@
+GSPORT_VERSION = "1.8.0"
